+++ tests/integration/reconcile/test_aggregates_reconcile.py
@@ -10,9 +10,10 @@ from pyspark.testing import assertDataFrameEqual
 from pyspark.sql import Row
 
 from databricks.labs.lakebridge.config import DatabaseConfig, ReconcileMetadataConfig
+from databricks.labs.lakebridge.reconcile.reconcilation import Reconciliation
 from databricks.labs.lakebridge.transpiler.sqlglot.dialect_utils import get_dialect
 from databricks.labs.lakebridge.reconcile.connectors.data_source import MockDataSource
-from databricks.labs.lakebridge.reconcile.execute import Reconciliation, main
+from databricks.labs.lakebridge.reconcile.execute import main
 from databricks.labs.lakebridge.reconcile.recon_config import (
     Aggregate,
     AggregateRule,
